requirements.txt
Django
django-cors-headers
django-ninja
django-ninja-jwt[crypto]
pydantic[email]
rav